backend/requirements.txt
# GhostCart Backend Dependencies
# Python 3.11+ required

# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Data Validation & Models
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
greenlet>=3.0.0  # Required for SQLAlchemy async support

# Scheduling
apscheduler>=3.10.0

# AWS Bedrock & Strands
boto3>=1.28.0
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0  # For testing FastAPI

# Code Quality
black>=23.0.0
mypy>=1.4.0
ruff>=0.0.280
